requirements.txt
IPython

Django
django-rest-framework


#router db
six

#image
Pillow